Added TradeRequest class and Card class w/ subclasses
A TradeRequest object holds the participants and proposed properties, having its own methods to carry out the trade if accepted. The Card class is meant to implement Community Chest and Chance functionality. It is, to compare to Tiles implementation, lacking a manager class and will be handled directly by the GameSystem. A psuedo-Command pattern is used in designing how the Cards work. There is opportunity for using dependency injection with how data is imported from databases for both Cards and Tiles.